Add Xml library
